Fix libsodium initialization
Libsodium was initialized later on in the program, after some functions
had already been called. This usually works, but can lead to surprising
problems. The initialization was moved to the start of the program.

Furthermore, the result value of sodium_init() was not correctly
checked, anything except 0 (success) was rejected, even though 1
(already initialized) should also be acceptable.